Waifu Selected Fully implemented
Works as intended and changes waifu sprite.

Also changes the text if the waifu its selected from Unlock to Select. And so it changes the behaivour of the button from removing points to changing the sprite in the game scene.

Notes: UI might need some increase in size, a bit on the small size. And yeah, get ART. pepesad